+++ tests/test_dcos_e2e/test_node.py
@@ -158,6 +158,48 @@ class TestSendFile:
         result = dcos_node.run(args=args)
         assert result.stdout.decode() == content
 
+    def test_send_directory(
+        self,
+        dcos_node: Node,
+        tmpdir: local,
+    ) -> None:
+        """
+        It is possible to send a directory to a cluster node as the default
+        user.
+        """
+        original_content = str(uuid.uuid4())
+        local_dir_name = 'local_dir'
+        remote_dir_name = 'remote_dir'
+        file_name = 'example_file.txt'
+        local_dir = tmpdir.mkdir(local_dir_name)
+        local_file = local_dir.join(file_name)
+        local_file.write(original_content)
+
+        random = uuid.uuid4().hex
+        master_base_dir = '/etc/{random}'.format(random=random)
+        master_destination_dir = Path(master_base_dir) / remote_dir_name
+
+        dcos_node.run(args=['mkdir', '-p', str(master_destination_dir)])
+        dcos_node.send_file(
+            local_path=Path(local_file),
+            remote_path=master_destination_dir / file_name,
+        )
+
+        args = ['cat', str(master_destination_dir / file_name)]
+        result = dcos_node.run(args=args)
+        assert result.stdout.decode() == original_content
+
+        new_content = str(uuid.uuid4())
+        local_file.write(new_content)
+
+        dcos_node.send_file(
+            local_path=Path(local_dir),
+            remote_path=master_destination_dir,
+        )
+        args = ['cat', str(master_destination_dir / file_name)]
+        result = dcos_node.run(args=args)
+        assert result.stdout.decode() == new_content
+
     def test_send_file_to_tmp_directory(
         self,
         dcos_node: Node,